Correcciones de errores en la sección de registro de pacientes, modificación de pacientes, eliminación de pacientes, corrección del módulo administrativo, esquematización de las funciones para reutilizar el código.

+++ trunk/sql/public/funciones/funciones del administrador/adm_modificar_medico.sql
@@ -4,6 +4,7 @@ $BODY$
 DECLARE
 	datos ALIAS 	FOR $1;
 	_id_doc		doctores.id_doc%TYPE;
+	_ced_doc	doctores.ced_doc%TYPE;
 	_nom_doc 	doctores.nom_doc%TYPE;
 	_ape_doc 	doctores.ape_doc%TYPE;
 	_pas_doc	doctores.pas_doc%TYPE;
@@ -19,11 +20,12 @@ BEGIN
 
 	_id_doc		:= datos[1];
 	_log_doc 	:= datos[2];
-	_nom_doc 	:= datos[3];
-	_ape_doc 	:= datos[4];
-	_pas_doc	:= md5(datos[5]);		
-	_tel_doc 	:= datos[6];
-	_trans_doc	:= datos[7];
+	_ced_doc	:= datos[3];
+	_nom_doc 	:= datos[4];
+	_ape_doc 	:= datos[5];
+	_pas_doc	:= md5(datos[6]);		
+	_tel_doc 	:= datos[7];
+	_trans_doc	:= datos[8];
 	
 	
 	IF EXISTS(SELECT 1 FROM doctores WHERE id_doc = _id_doc)THEN
@@ -31,38 +33,44 @@ BEGIN
 		IF NOT EXISTS (SELECT 1 FROM doctores WHERE log_doc = _log_doc AND id_doc <> _id_doc) THEN     					
 			/*Inserta registro en la tabla usuarios_administrativos*/	
 
-			UPDATE doctores SET 
-				nom_doc = _nom_doc,
-				ape_doc = _ape_doc,
-				pas_doc = _pas_doc,
-				log_doc = _log_doc,
-				tel_doc = _tel_doc					
-			
-			WHERE 
-			id_doc = _id_doc
-			;
-			
-			/* Insertando las transacciones del usuario*/
-			_id_tip_usu_usu := (SELECT id_tip_usu_usu FROM tipos_usuarios__usuarios WHERE id_doc = _id_doc);
-			DELETE FROM transacciones_usuarios WHERE id_tip_usu_usu = _id_tip_usu_usu;	
-			
-			_arr_trans_doc := STRING_TO_ARRAY(_trans_doc,',');
+			IF NOT EXISTS(SELECT 1 FROM doctores WHERE ced_doc = _ced_doc AND id_doc <> _id_doc) THEN     					
+				UPDATE doctores SET 
 					
-			IF (ARRAY_UPPER(_arr_trans_doc,1) > 0)THEN							
-				FOR i IN 1..(ARRAY_UPPER(_arr_trans_doc,1)) LOOP
-					INSERT INTO transacciones_usuarios(
-						id_tip_usu_usu,
-						id_tip_tra
-					)
-					VALUES
-					(					
-						_id_tip_usu_usu,
-						_arr_trans_doc[i]
-					);
-				END LOOP;
+					nom_doc = _nom_doc,
+					ape_doc = _ape_doc,
+					pas_doc = _pas_doc,
+					ced_doc	= _ced_doc,
+					log_doc = _log_doc,
+					tel_doc = _tel_doc					
+				
+				WHERE 
+				id_doc = _id_doc
+				;
+				
+				/* Insertando las transacciones del usuario*/
+				_id_tip_usu_usu := (SELECT id_tip_usu_usu FROM tipos_usuarios__usuarios WHERE id_doc = _id_doc);
+				DELETE FROM transacciones_usuarios WHERE id_tip_usu_usu = _id_tip_usu_usu;	
+				
+				_arr_trans_doc := STRING_TO_ARRAY(_trans_doc,',');
+						
+				IF (ARRAY_UPPER(_arr_trans_doc,1) > 0)THEN							
+					FOR i IN 1..(ARRAY_UPPER(_arr_trans_doc,1)) LOOP
+						INSERT INTO transacciones_usuarios(
+							id_tip_usu_usu,
+							id_tip_tra
+						)
+						VALUES
+						(					
+							_id_tip_usu_usu,
+							_arr_trans_doc[i]
+						);
+					END LOOP;
+				END IF;
+			ELSE
+				RETURN 3;
 			END IF;
 
-			-- La función se ejecutó exitosamente
+				-- La función se ejecutó exitosamente
 			RETURN 1;
 			
 		
@@ -93,8 +101,12 @@ DESCRIPCION:
 	Almacena la información del doctor
 
 RETORNO:
-	1: La función se ejecutó exitosamente
+
 	0: Existe un usuario administrativo con el mismo login
+	1: La función se ejecutó exitosamente
+	2: Ya existe un usuario administrativo con este login
+	3: Ya existe un usuario administrativo con la misma cédula
+	
 	 
 EJEMPLO DE LLAMADA:
 	SELECT adm_modificar_medico(ARRAY[''1'',''Lisseth'', ''Lozada'', ''123'', ''llozada'',''04269150722'',''1,2,3'']);
